refactor : 일급 객체 하나로 묶기
