+++ src/main/java/nextstep/subway/applicaion/LineService.java
@@ -5,6 +5,7 @@ import nextstep.subway.applicaion.dto.LineResponse;
 import nextstep.subway.applicaion.dto.SectionRequest;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.LineRepository;
+import nextstep.subway.domain.Section;
 import nextstep.subway.domain.Station;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -68,7 +69,7 @@ public class LineService {
         Station downStation = stationService.findById(sectionRequest.getDownStationId());
         Line line = findById(lineId);
 
-        line.addSection(upStation, downStation, sectionRequest.getDistance(),sectionRequest.getDuration());
+        line.addSection(upStation, downStation, new Section.SectionInfo(sectionRequest.getDistance(), sectionRequest.getDuration()));
     }
 
     public void deleteSection(Long lineId, Long stationId) {
