--- src/main/java/nextstep/subway/domain/Duration.java
@@ -1,17 +1,0 @@
-package nextstep.subway.domain;
-
-import javax.persistence.Embeddable;
-
-@Embeddable
-public class Duration {
-
-    private int duration;
-
-    public Duration(int duration) {
-        this.duration = duration;
-    }
-
-    public int getDuration() {
-        return duration;
-    }
-}
